Add transparency change and allow resizing window
When you mouse over a square, it goes from fully transparent to a random
color. When you mouse over it again, it will keep the color but become
less transparent (10% each time).

Also, when resizing the browser window, the viewable square changes
size as well (including all the smaller squares). Colors are not
affected by resizing.